Refactor: move agents into separate dir
