+++ agents/crl/crl.py
@@ -8,127 +8,22 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 import optax
-import wandb_osh
 from brax import base, envs
 from brax.training import types
 from brax.v1 import envs as envs_v1
 from etils import epath
-from flax.linen.initializers import variance_scaling
 from flax.struct import dataclass
 from flax.training.train_state import TrainState
-from wandb_osh.hooks import TriggerWandbSyncHook
 
-import wandb
 from envs.wrappers import TrajectoryIdWrapper
-from src.evaluator import CrlEvaluator
-from src.replay_buffer import TrajectoryUniformSamplingQueue
-from utils import render_policy
+from utils.evaluator import CrlEvaluator
+from utils.replay_buffer import TrajectoryUniformSamplingQueue
+
+from .networks import Actor, Encoder
 
 Metrics = types.Metrics
 Env = Union[envs.Env, envs_v1.Env, envs_v1.Wrapper]
 State = Union[envs.State, envs_v1.State]
-
-
-class Encoder(nn.Module):
-    repr_dim: int = 64
-    network_width: int = 256
-    network_depth: int = 4
-    skip_connections: int = (
-        0  # 0 for no skip connections, >= 0 means the frequency of skip connections (every X layers)
-    )
-    use_relu: bool = False
-    use_ln: bool = False
-
-    @nn.compact
-    def __call__(self, data: jnp.ndarray):
-
-        lecun_unfirom = variance_scaling(1 / 3, "fan_in", "uniform")
-        bias_init = nn.initializers.zeros
-
-        if self.use_ln:
-            normalize = lambda x: nn.LayerNorm()(x)
-        else:
-            normalize = lambda x: x
-
-        if self.use_relu:
-            activation = nn.relu
-        else:
-            activation = nn.swish
-
-        x = data
-        for i in range(self.network_depth):
-            x = nn.Dense(
-                self.network_width, kernel_init=lecun_unfirom, bias_init=bias_init
-            )(x)
-            x = normalize(x)
-            x = activation(x)
-
-            if self.skip_connections:
-                if i == 0:
-                    skip = x
-                if i > 0 and i % self.skip_connections == 0:
-                    x = x + skip
-                    skip = x
-
-        x = nn.Dense(self.repr_dim, kernel_init=lecun_unfirom, bias_init=bias_init)(x)
-        return x
-
-
-class Actor(nn.Module):
-    action_size: int
-    network_width: int = 256
-    network_depth: int = 4
-    skip_connections: int = (
-        0  # 0 for no skip connections, >= 0 means the frequency of skip connections (every X layers)
-    )
-    use_relu: bool = False
-    use_ln: bool = False
-    LOG_STD_MAX = 2
-    LOG_STD_MIN = -5
-
-    @nn.compact
-    def __call__(self, x):
-        if self.use_ln:
-            normalize = lambda x: nn.LayerNorm()(x)
-        else:
-            normalize = lambda x: x
-
-        if self.use_relu:
-            activation = nn.relu
-        else:
-            activation = nn.swish
-
-        lecun_unfirom = variance_scaling(1 / 3, "fan_in", "uniform")
-        bias_init = nn.initializers.zeros
-
-        # print(f"x.shape: {x.shape}", flush=True)
-        for i in range(self.network_depth):
-            x = nn.Dense(
-                self.network_width, kernel_init=lecun_unfirom, bias_init=bias_init
-            )(x)
-            x = normalize(x)
-            x = activation(x)
-
-            if self.skip_connections:
-                if i == 0:
-                    skip = x
-                if i > 0 and i % self.skip_connections == 0:
-                    x = x + skip
-                    skip = x
-
-        mean = nn.Dense(
-            self.action_size, kernel_init=lecun_unfirom, bias_init=bias_init
-        )(x)
-        log_std = nn.Dense(
-            self.action_size, kernel_init=lecun_unfirom, bias_init=bias_init
-        )(x)
-
-        log_std = nn.tanh(log_std)
-        log_std = self.LOG_STD_MIN + 0.5 * (self.LOG_STD_MAX - self.LOG_STD_MIN) * (
-            log_std + 1
-        )  # From SpinUp / Denis Yarats
-
-        return mean, log_std
 
 
 @dataclass
@@ -194,9 +89,9 @@ class CRL:
             NOTE: `num_envs * (episode_length - 1)` must be divisible by
             `batch_size` due to the way data is stored in replay buffer.
         """
-        assert config.num_envs * (config.episode_length - 1) % self.batch_size == 0, (
-            "num_envs * (episode_length - 1) must be divisible by batch_size"
-        )
+        assert (
+            config.num_envs * (config.episode_length - 1) % self.batch_size == 0
+        ), "num_envs * (episode_length - 1) must be divisible by batch_size"
 
     def train_fn(
         self,
@@ -209,7 +104,6 @@ class CRL:
         progress_fn: Callable[[int, Metrics], None] = lambda *args: None,
         checkpoint_logdir: Optional[str] = None,
     ):
-
 
         self.check_config(config)
 
@@ -733,7 +627,6 @@ class CRL:
                 do_render=do_render,
             )
 
-
             if config.checkpoint:
                 # Save current policy and critic params.
                 params = (
@@ -744,7 +637,6 @@ class CRL:
                 path = f"{config.ckpt_dir}/step_{int(training_state.env_steps)}.pkl"
                 save_params(path, params)
 
-
         if config.checkpoint:
             # Save current policy and critic params.
             params = (
@@ -754,4 +646,3 @@ class CRL:
             )
             path = f"{config.ckpt_dir}/final.pkl"
             save_params(path, params)
-
