+++ agents/td3/losses.py
@@ -9,7 +9,7 @@ import jax.numpy as jnp
 from brax.training import types
 from brax.training.types import Params, PRNGKey
 
-from src.td3 import networks
+from agents.td3 import networks
 
 Transition = types.Transition
