--- src/td3/__init__.py
@@ -1,1 +1,0 @@
-from .td3 import TD3
